[bug 1139713] Add received_ts field to hb Answer model
* add a received_ts field to the hb Answer model with a silly
  initial default value
* backfill the field with the data from updated_ts
